requirements.txt
# PDEBench稀疏观测重建系统 - 依赖包列表

# 核心深度学习框架
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# 数值计算和科学计算
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# 图像处理
pillow>=9.0.0
scikit-image>=0.20.0
opencv-python>=4.7.0

# 配置管理
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# 数据处理
h5py>=3.8.0
netcdf4>=1.6.0
xarray>=2023.1.0

# 机器学习工具
scikit-learn>=1.2.0

# 进度条和日志
tqdm>=4.64.0
tensorboard>=2.12.0
wandb>=0.15.0

# 开发工具
pytest>=7.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# 系统监控
psutil>=5.9.0

# Jupyter支持（可选）
jupyter>=1.0.0
ipywidgets>=8.0.0

# 其他工具
requests>=2.28.0
urllib3>=1.26.0